Quash miscellaneous compiler warnings.
Missing #include for strcmp().
Using char for array index.
Change (A || B && C) to (A || (B && C)) for clarity.
Remove declaration of unused variable.
Add missing return type for main().

+++ extract.c
@@ -88,7 +88,7 @@ int main
        block boundary. */
 
     for (i = M; i<N; i++)
-    { putc("01"[cblk[cols[i]]],extf);
+    { putc("01"[(int)cblk[cols[i]]],extf);
     }
    
     putc('\n',extf);
